+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.27](https://github.com/arez/arez/tree/v0.27) (2017-11-28)
 [Full Changelog](https://github.com/arez/arez/compare/v0.26...v0.27)